refactor: Hit, Started, Running을 Running 하나로 통합
