--- src/main/java/blackjack/domain/state/Hit.java
@@ -1,10 +1,0 @@
-package blackjack.domain.state;
-
-import blackjack.domain.card.Cards;
-
-public class Hit extends Running {
-
-    Hit(Cards cards) {
-        super(cards);
-    }
-}
